+++ client/src/components/inputs/Label/Label.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const Label: React.FC = () => {
+  return <>Label</>;
+};
+
+export default Label;
