+++ client/src/hook/use-drop-down/index.tsx
@@ -1,4 +1,4 @@
-import { Options } from '../../components/inputs/Select';
+import { Options } from '../../components/inputs/Select/Select';
 
 const isString = (target: string | string[] | undefined): target is string => {
   return typeof target == 'string';
